Rename namespace Adyen.CloudApiSerialization to Adyen.Serialization
