+++ Adyen/Security/SaleToPoiMessageSecured.cs
@@ -21,7 +21,7 @@
 //  */
 #endregion
 
-using Adyen.CloudApiSerialization;
+using Adyen.Serialization;
 using Adyen.Model.Nexo;
 using Adyen.Model.Nexo.Message;
