+++ Adyen/Serialization/Converter/JsonBase64Converter.cs
@@ -20,29 +20,32 @@
 //  * See the LICENSE file for more info.
 //  */
 #endregion
-
-using Adyen.Exceptions;
 using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
+using System;
 
-namespace Adyen.CloudApiSerialization
+namespace Adyen.Serialization.Converter
 {
-    internal class JSonConvertDeserializerWrapper<T>
+    internal class JsonBase64Converter : JsonConverter
     {
-        internal static T DeserializeObject(string objectToDeserialize)
+        public override bool CanConvert(Type objectType)
         {
-            return JsonConvert.DeserializeObject<T>(objectToDeserialize, new JsonSerializerSettings
-            {
-                MissingMemberHandling = MissingMemberHandling.Error,
-                Error = delegate (object sender, ErrorEventArgs args)
-                {
-                    var exceptionMessage = string.Format(ExceptionMessages.ExceptionDuringDeserialization,
-                        objectToDeserialize,
-                        args.ErrorContext.Error.Message);
+            return true;
+        }
 
-                    throw new DeserializationException(exceptionMessage, args.ErrorContext.Error);
-                }
-            });
+        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
+        {
+            return System.Text.Encoding.UTF8.GetString((Convert.FromBase64String((string)reader.Value)));
+        }
+
+        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
+        {
+            var json = JsonConvert.SerializeObject(value, Formatting.None,
+                            new JsonSerializerSettings
+                            {
+                                NullValueHandling = NullValueHandling.Ignore
+                            });
+
+            writer.WriteValue(Convert.ToBase64String(System.Text.Encoding.UTF8.GetBytes(json)));
         }
     }
 }
