+++ Adyen/Serialization/IMessagePayload.cs
@@ -21,14 +21,7 @@
 //  */
 #endregion
 
-namespace Adyen.CloudApiSerialization
+namespace Adyen.Serialization
 {
-    internal class MessagePayloadSerializer<T> : IMessagePayloadSerializer<T> where T : IMessagePayload
-    {
-        public IMessagePayload Deserialize(string messagePayloadJson)
-        {
-            return Converter.JSonConvertDeserializerWrapper<T>.DeserializeObject(messagePayloadJson);
-        }
-
-    }
+    public interface IMessagePayload { }
 }
