+++ Adyen/Serialization/IMessagePayloadSerializer.cs
@@ -20,32 +20,11 @@
 //  * See the LICENSE file for more info.
 //  */
 #endregion
-using Newtonsoft.Json;
-using System;
 
-namespace Adyen.CloudApiSerialization.Converter
+namespace Adyen.Serialization
 {
-    internal class JsonBase64Converter : JsonConverter
+    internal interface IMessagePayloadSerializer<out T> where T : IMessagePayload
     {
-        public override bool CanConvert(Type objectType)
-        {
-            return true;
-        }
-
-        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
-        {
-            return System.Text.Encoding.UTF8.GetString((Convert.FromBase64String((string)reader.Value)));
-        }
-
-        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
-        {
-            var json = JsonConvert.SerializeObject(value, Formatting.None,
-                            new JsonSerializerSettings
-                            {
-                                NullValueHandling = NullValueHandling.Ignore
-                            });
-
-            writer.WriteValue(Convert.ToBase64String(System.Text.Encoding.UTF8.GetBytes(json)));
-        }
+        IMessagePayload Deserialize(string messagePayloadJson);
     }
 }
