+++ Adyen/Serialization/MessagePayloadSerializer.cs
@@ -21,7 +21,14 @@
 //  */
 #endregion
 
-namespace Adyen.CloudApiSerialization
+namespace Adyen.Serialization
 {
-    public interface IMessagePayload { }
+    internal class MessagePayloadSerializer<T> : IMessagePayloadSerializer<T> where T : IMessagePayload
+    {
+        public IMessagePayload Deserialize(string messagePayloadJson)
+        {
+            return Converter.JSonConvertDeserializerWrapper<T>.DeserializeObject(messagePayloadJson);
+        }
+
+    }
 }
