+++ Adyen/Serialization/SaleToPoiMessageSerializer.cs
@@ -25,7 +25,7 @@ using Adyen.Model.Nexo;
 using Adyen.Security;
 using Newtonsoft.Json.Linq;
 
-namespace Adyen.CloudApiSerialization
+namespace Adyen.Serialization
 {
     public class SaleToPoiMessageSerializer
     {
